Appコンポーネントのリファクタリング
・Eventsコンポーネントの実装
・EventFormコンポーネントの実装
・Eventコンポーネントのリファクタリング
